requirements.txt
PyQt6==6.11.0
umap-learn==0.5.12
hdbscan==0.8.42
ruff==0.15.12